agent-sandbox: split the dispatcher into one file per boundary
No behaviour change. dispatcher.py becomes four modules, imports running
strictly downward, each named for a boundary the process crosses:

    __main__  -> http_api -> tasks -> kube

  kube      the Kubernetes API and the Job template — the task-pod security
            boundary, reviewable on its own
  tasks     the task table and the run-one-Job-to-completion loop
  http_api  parse, bound, dispatch, answer
  __main__  wiring, and nothing else

The split is preemptive: OIDC verification and caller authorization are landing
in this process next, and a single file that already mixes the HTTP surface with
the code holding create-Job RBAC is the wrong thing to add a JWT parser to.

Two mechanical notes. Modules import modules and call through them
(kube.job_state(...), not `from kube import job_state`), so a test substituting
a call reaches the one binding that exists. And the container entrypoint is now
`python3 <dir>`, which runs __main__.py with <dir> on sys.path — no package, no
PYTHONPATH, no install step.

Test plan: the same 55 tests, ported to the new module names and passing
unchanged; the entrypoint additionally smoke-tested the way the container
invokes it (python3 <dir> -> listening on [::], /healthz 200, unknown path 404).
